# internal/api/thea.openapi.yaml
openapi: 3.0.0
info:
  title: Thea Spec
  description: "Thea REST API spec. See http://github.com/hbomb79/Thea for more information"
  version: 1.0.0
  contact:
    name: Thea Support
    url: https://github.com/hbomb79/Thea
...
security:
  - permissionAuth: [] # Default security - requires authentication but no specific permissions
paths:
...
  /transcodes:
    post:
      summary: Create a new transcode task
      description: Creates a new task
      operationId: createTranscodeTask
      tags:
        - Transcode Tasks
      security:
        - permissionAuth: [transcode:create]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTranscodeTaskRequest"
      responses:
        "201":
          description: Creation successful
  /transcodes/active:
    get:
      summary: List Active Tasks
      description: Returns all active tasks
      operationId: listActiveTranscodeTasks
      tags:
        - Transcode Tasks
      security:
        - permissionAuth: [transcode:access]
      responses:
        "200":
          description: List of active tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TranscodeTask"
  /transcodes/complete:
    get:
      summary: List Completed Tasks
      description: Returns all completed tasks
      operationId: listCompletedTranscodeTasks
      tags:
        - Transcode Tasks
      security:
        - permissionAuth: [transcode:access]
      responses:
        "200":
          description: List of complete tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TranscodeTask"
  /transcodes/{id}:
    get:
      summary: Get Transcode Task
      description: Return matching task, whether active or completed
      operationId: getTranscodeTask
      tags:
        - Transcode Tasks
      security:
        - permissionAuth: [transcode:access]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Complete or active transcode task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscodeTask"
    delete:
      summary: Delete Task
      description: Deletes matching task. Cancells it if active, and cleans up transcoded media if completed.
      operationId: deleteTranscodeTask
      tags:
        - Transcode Tasks
      security:
        - permissionAuth: [transcode:access, transcode:delete]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Delete successful
  /transcodes/{id}/pause:
    post:
      summary: Pause Task
      description: Pauses an ffmpeg task, halting it's progress
      operationId: pauseTranscodeTask
      parameters:
        - $ref: "#/components/parameters/ID"
      tags:
        - Transcode Tasks
      security:
        - permissionAuth: [transcode:access, transcode:modify]
      responses:
        "200":
          description: Transcode paused
  /transcodes/{id}/resume:
    post:
      summary: Resume Task
      description: Resumes an already paused ffmpeg task
      operationId: resumeTranscodeTask
      tags:
        - Transcode Tasks
      security:
        - permissionAuth: [transcode:access, transcode:modify]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Transcode resumed
...
  /transcode-targets:
    get:
      tags:
        - Targets
      security:
        - permissionAuth: [target:access]
      summary: List Targets
      description: Returns all known targets
      operationId: listTargets
      responses:
        "200":
          description: A list of the targets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Target"
    post:
      tags:
        - Targets
      security:
        - permissionAuth: [target:create]
      summary: Create Target
      description: Creates a new target
      operationId: createTarget
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTargetRequest"
      responses:
        "201":
          description: The created target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Target"
        "400":
          description: Invalid request
  /transcode-targets/{id}:
    get:
      tags:
        - Targets
      security:
        - permissionAuth: [target:access]
      summary: Get Target
      description: Fetches the target with the ID specified
      operationId: getTarget
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: The target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Target"
        "400":
          description: Invalid request
    patch:
      tags:
        - Targets
      security:
        - permissionAuth: [target:access, target:modify]
      summary: Update Target
      description: Updates the target specified using the ID
      operationId: updateTarget
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTargetRequest"
      responses:
        "200":
          description: The updated target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Target"
        "400":
          description: Invalid request
    delete:
      tags:
        - Targets
      security:
        - permissionAuth: [target:access, target:delete]
      summary: Delete Target
      description: Delete the target specified using the ID
      operationId: deleteTarget
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Delete success
...
servers:
  - url: http://localhost:8080/api/thea/v1
components:
  securitySchemes:
    permissionAuth:
      type: apiKey
      in: cookie
      name: auth-token
  parameters:
    ID:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    CreateTranscodeTaskRequest:
      type: object
      required:
        - media_id
        - target_id
      properties:
        media_id:
          type: string
          format: uuid
        target_id:
          type: string
          format: uuid
    TranscodeTaskStatus:
      type: string
      enum: ['WAITING', 'WORKING', 'SUSPENDED', 'TROUBLED', 'CANCELLED', 'COMPLETE']
    TranscodeTaskProgress:
      type: object
      required:
        - frames_processed
        - current_time
        - current_bitrate
        - progress
        - speed
      properties:
        frames_processed:
          type: string
        current_time:
          type: string
        current_bitrate:
          type: string
        progress:
          type: number
        speed:
          type: string
    TranscodeTask:
      type: object
      required:
        - id
        - media_id
        - target_id
        - output_path
        - status
      properties:
        id:
          type: string
          format: uuid
        media_id:
          type: string
          format: uuid
        target_id:
          type: string
          format: uuid
        output_path:
          type: string
        status:
          $ref: "#/components/schemas/TranscodeTaskStatus"
        progress:
          $ref: "#/components/schemas/TranscodeTaskProgress"
...
    Target:
      type: object
      required:
        - id
        - label
        - extension
        - ffmpeg_options
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        extension:
          type: string
        ffmpeg_options:
          type: object
    CreateTargetRequest:
      type: object
      required:
        - label
        - extension
        - ffmpeg_options
      properties:
        label:
          type: string
        extension:
          type: string
        ffmpeg_options:
          type: object
    UpdateTargetRequest:
      type: object
      properties:
        label:
          type: string
        extension:
          type: string
        ffmpeg_options:
          type: object